# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DOCX Exporter REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/convert": {
      "post": {
        "operationId": "convertToDocx",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportParams"
              }
            }
          },
          "description": "Export parameters to generate the DOCX",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {}
            },
            "description": "Content of DOCX document as a byte array",
            "headers": {
              "Content-Disposition": {
                "description": "To inform a browser that the response is a downloadable attachment",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              },
              "Export-Filename": {
                "description": "File name for converted DOCX document",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              },
              "Missing-WorkItem-Attachments-Count": {
                "description": "Unavailable work item attachments count",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              },
              "WorkItem-IDs-With-Missing-Attachment": {
                "description": "Work items contained unavailable attachments",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          }
        },
        "summary": "Returns requested Polarion's document converted to DOCX",
        "tags": [
          "DOCX Processing"
        ]
      }
    },
    "/api/convert/html": {
      "post": {
        "operationId": "convertHtmlToPdf",
        "parameters": [
          {
            "description": "default value: document.docx",
            "in": "query",
            "name": "fileName",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "html": {
                    "description": "Required html file",
                    "format": "binary",
                    "type": "string"
                  },
                  "params": {
                    "description": "Optional params json object",
                    "type": "string"
                  },
                  "template": {
                    "description": "Optional template file",
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "html"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {}
            },
            "description": "Content of DOCX document as a byte array",
            "headers": {
              "Content-Disposition": {
                "description": "To inform a browser that the response is a downloadable attachment",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              },
              "Export-Filename": {
                "description": "File name for converted DOCX document",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          }
        },
        "summary": "Converts input HTML to DOCX",
        "tags": [
          "DOCX Processing"
        ]
      }
    },
    "/api/convert/jobs": {
      "get": {
        "operationId": "getAllPdfConverterJobs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Conversion jobs statuses"
          }
        },
        "summary": "Returns all active DOCX conversion jobs statuses",
        "tags": [
          "DOCX Processing"
        ]
      },
      "post": {
        "operationId": "startPdfConverterJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportParams"
              }
            }
          },
          "description": "Export parameters to generate the DOCX",
          "required": true
        },
        "responses": {
          "202": {
            "description": "Conversion process is started, job URI is returned in Location header"
          }
        },
        "summary": "Starts asynchronous conversion job of Polarion's document to DOCX",
        "tags": [
          "DOCX Processing"
        ]
      }
    },
    "/api/convert/jobs/{id}": {
      "get": {
        "operationId": "getPdfConverterJobStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConverterJobDetails"
                }
              }
            },
            "description": "Conversion job is still in progress"
          },
          "303": {
            "content": {
              "application/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConverterJobDetails"
                }
              }
            },
            "description": "Conversion job is finished successfully, Location header contains result URL"
          },
          "404": {
            "description": "Conversion job id is unknown"
          },
          "409": {
            "description": "Conversion job is failed or cancelled"
          }
        },
        "summary": "Returns DOCX conversion job status",
        "tags": [
          "DOCX Processing"
        ]
      }
    },
    "/api/convert/jobs/{id}/result": {
      "get": {
        "operationId": "getPdfConverterJobResult",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {}
            },
            "description": "Conversion job result is ready",
            "headers": {
              "Content-Disposition": {
                "description": "To inform a browser that the response is a downloadable attachment",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              },
              "Export-Filename": {
                "description": "File name for converted DOCX document",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              },
              "Missing-WorkItem-Attachments-Count": {
                "description": "Unavailable work item attachments count",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              },
              "WorkItem-IDs-With-Missing-Attachment": {
                "description": "Work items contained unavailable attachments",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "204": {
            "description": "Conversion job is still in progress"
          },
          "404": {
            "description": "Conversion job id is unknown"
          },
          "409": {
            "description": "Conversion job is failed, cancelled or result is unreachable"
          }
        },
        "summary": "Returns DOCX conversion job result",
        "tags": [
          "DOCX Processing"
        ]
      }
    },
...
    "/api/export-filename": {
      "post": {
        "operationId": "getFileName",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportParams"
              }
            }
          },
          "description": "Export parameters similar as for generation of PDF",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully generated filename"
          }
        },
        "summary": "Gets a filename, prepared with velocity and placeholders",
        "tags": [
          "Utility resources"
        ]
      }
    },
...
    "/api/prepared-html-content": {
      "post": {
        "operationId": "prepareHtmlContent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportParams"
              }
            }
          },
          "description": "Export parameters to generate the DOCX",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/html": {}
            },
            "description": "Prepared HTML content"
          }
        },
        "summary": "Returns prepared HTML which will be used for DOCX conversion using Pandoc",
        "tags": [
          "DOCX Processing"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConverterJobDetails": {
        "description": "Details of the converter job including status and error message if any",
        "properties": {
          "errorMessage": {
            "description": "Error message if the conversion failed",
            "type": "string"
          },
          "status": {
            "description": "Status of the converter job",
            "enum": [
              "IN_PROGRESS",
              "SUCCESSFULLY_FINISHED",
              "FAILED",
              "CANCELLED"
            ],
            "example": "IN_PROGRESS",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
...
}